pyproject.toml
[tool.poetry]
name = "blenderaddoninstaller"
version = "0.1.0"
description = ""
authors = ["Christopher Winkelmann <winkelchri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
pyyaml = "^5.1"
autopep8 = "^1.4"
loguru = "^0.3.2"
jsonschema = "^3.0"

[tool.poetry.dev-dependencies]
pep8 = "^1.7"
pylint = "^2.3"
autopep8 = "^1.4"
pytest = "^5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
